# Optimize JAX performance in data preparation pipeline