--- a/tests/interface_dag_elements/test_failures.py
+++ b/tests/interface_dag_elements/test_failures.py
@@ -890,7 +890,7 @@ def test_fail_if_p_id_does_not_exist(xnp):
         ValueError,
         match="The input data must contain the `p_id` column.",
     ):
-        input_data_is_invalid(data)
+        input_data_is_invalid(data, xnp)
 
 
 def test_fail_if_p_id_is_missing_via_main(backend):
@@ -917,7 +917,7 @@ def test_fail_if_p_id_is_not_unique(xnp):
         ValueError,
         match="The following `p_id`s are not unique in the input data",
     ):
-        input_data_is_invalid(data)
+        input_data_is_invalid(data, xnp)
 
 
 def test_fail_if_p_id_is_not_unique_via_main(minimal_input_data, backend):
@@ -948,12 +948,12 @@ def test_fail_if_p_id_is_not_unique_via_main(minimal_input_data, backend):
         {("p_id",): pd.Series([1, "2", 3.0])},
     ],
 )
-def test_fail_if_p_id_is_not_int(data):
+def test_fail_if_p_id_is_not_int(data, xnp):
     with pytest.raises(
         ValueError,
         match="The `p_id` column must be of integer dtype.",
     ):
-        input_data_is_invalid(data)
+        input_data_is_invalid(data, xnp)
 
 
 @pytest.mark.parametrize(
@@ -963,14 +963,32 @@ def test_fail_if_p_id_is_not_int(data):
         {("p_id",): pd.Series([1, 2, 3])},
     ],
 )
-def test_p_id_can_be_specified_as_series_and_numpy_array(data):
-    input_data_is_invalid(data)
+def test_p_id_can_be_specified_as_series_and_numpy_array(data, xnp):
+    input_data_is_invalid(data, xnp)
 
 
 @pytest.mark.skipif_numpy
 def test_p_id_can_be_specified_as_jax_array(xnp):
     data = {("p_id",): xnp.array([1, 2, 3])}
-    input_data_is_invalid(data)
+    input_data_is_invalid(data, xnp)
+
+
+def test_single_person_with_p_id_zero_does_not_trigger_duplicate_error(xnp):
+    """Test edge case: single person with p_id=0 should not be flagged as duplicate.
+
+    Without the single-row protection in input_data_is_invalid, this test would fail
+    because the duplicate detection logic incorrectly identifies p_id=0 as a duplicate
+    when using xnp.diff([0], append=0) which gives [0], and 0 == 0 is True.
+    """
+    data = {("p_id",): xnp.array([0])}
+    input_data_is_invalid(data, xnp)
+
+
+def test_single_person_with_any_p_id_works_correctly(xnp):
+    """Test that single-row data works for any p_id value."""
+    for p_id_value in [-100, 0, 1, 42, 999]:
+        data = {("p_id",): xnp.array([p_id_value])}
+        input_data_is_invalid(data, xnp)
 
 
 def test_fail_if_input_data_has_different_lengths(backend):