Let me start by letting the numbers speak for themselves:

```python
Benchmark Comparison Report
Generated: 2025-08-06T15:13:33.129062
Main branch file: benchmark_results_20250806_150436.json
PR branch file: benchmark_results_20250806_150643.json

==========================================================================================
JAX BACKEND COMPARISON: Main Branch vs PR Branch
==========================================================================================
Households  JAX hash main  JAX hash PR    JAX main (s)   JAX PR (s)     Speedup     
------------------------------------------------------------------------------------------
32,767      9a97a0ef       9a97a0ef       2.4546         0.7370         3.33x       
32,768      c0a7c7e9       c0a7c7e9       2.0527         0.5846         3.51x       
65,536      f873c7c8       f873c7c8       3.4010         0.6154         5.53x       
131,072     84608262       84608262       6.2885         0.7180         8.76x       
262,144     2fb136ba       2fb136ba       12.1383        1.0488         11.57x      
524,288     dd61a524       dd61a524       23.6647        1.9196         12.33x      
1,048,576   6ed4af0a       6ed4af0a       46.8753        3.8041         12.32x      
2,097,152   ff9f600f       ff9f600f       96.5656        8.0620         11.98x      

==========================================================================================
NUMPY BACKEND COMPARISON: Main Branch vs PR Branch
==========================================================================================
Households  NP hash main   NP hash PR     NP main (s)    NP PR (s)      Speedup     
------------------------------------------------------------------------------------------
32,767      2584ff76       2584ff76       0.1028         0.0907         1.13x       
32,768      a4ec6c3f       a4ec6c3f       0.1031         0.0692         1.49x       
65,536      e18ee207       e18ee207       0.1767         0.1279         1.38x       
131,072     e949fd41       e949fd41       0.4076         0.2277         1.79x       
262,144     a8fe5730       a8fe5730       0.7846         0.4868         1.61x       
524,288     c7c95469       c7c95469       1.3177         0.9271         1.42x       
1,048,576   1fbbd4ff       1fbbd4ff       2.8963         2.1271         1.36x       
2,097,152   81209a75       81209a75       6.0628         4.9379         1.23x       

======================================================================
SUMMARY STATISTICS
======================================================================

NUMPY Backend:
--------------------
  Average speedup: 1.43x
  Maximum speedup: 1.79x
  Minimum speedup: 1.13x
  Successful runs: 8/8
  Hash mismatches: 0/8
  ✓ All results are numerically identical

JAX Backend:
--------------------
  Average speedup: 8.67x
  Maximum speedup: 12.33x
  Minimum speedup: 3.33x
  Successful runs: 8/8
  Hash mismatches: 0/8
  ✓ All results are numerically identical
```

### Package for reproducing the benchmark (thanks, Claude): [benchmark.zip](https://github.com/user-attachments/files/21620776/benchmark.zip)

### Explanation/Remarks: 

- Profiling TTSIM (`main(main_targets=MainTarget.processed_data,...)`) for large datasets with the JAX backend shows that the majority of the runtime is spent in `data_converers.df_with_mapped_columns_to_flat_data` and `fail_if.input_data_is_invalid`
  
  ![gettsim_jax_1_000_000_main](https://github.com/user-attachments/assets/2bb7c6e9-6f7c-4dbb-880e-2f7b1f0d8d65)

- Regarding `data_converers.df_with_mapped_columns_to_flat_data`: For the JAX backend, it is much faster to create NumPy arrays first and then convert them to JAX arrays rather than to immediately create JAX arrays.
- Regarding `fail_if.input_data_is_invalid`: The loop over the JAX array is very slow. I think it should be safe to convert `p_id` to a NumPy array and then use Pandas for the duplicates detection.
- I expected the NumPy backend to be more or less unaffected, but it got a little faster as well...
- In case you are wondering "Why 32,767 and 32,768 households?" I found out by experimenting that `xnp.asarray(numpy_array, copy=False)` still copies for arrays smaller than 1MB; at exactly 1MB, JAX gives up and stops creating copies. The duplicated example household has four members, thus 32,768 * 4 * 8 = 1,048,567 = 1 MB (for a 64 bit int or float column). In the results you can see that 32.767 is (consistently, I checked) slower than 32.768 probably because in the former case arrays still get copied despite setting `xnp.asarray(numpy_array, copy=False)`.
- 🔪 ☠️ : You can use the `copy=False` option to mutate "immutable" JAX arrays by mutating the NumPy array that shares the same memory (if the array is >=1MB)! 😮There was probably a heated debate about this "feature" among the JAX developers...
- After the two optimizations the profile looks like this. The bottlenecks are gone.
  
  ![gettsim_jax_1_000_000_PR](https://github.com/user-attachments/assets/ad29f4e0-93a1-42dc-a7f0-046c54020828)

### TODOs

- [x] I am not really sure if this requires new tests. The code just replaces existing code.
- [x] I would like to know if this also results in better performance for JAX-CUDA and/or Metal, but I don't have the hardware to check this...