@mj023 2025-08-07
Old Code on GPU:

```
============================================================
PERFORMANCE COMPARISON
============================================================
Households  NumPy (s)   JAX (s)     NumPy Hash  JAX Hash    Speedup     
------------------------------------------------------------------------
32,767      0.0514      6.1766      2584ff76    9a97a0ef    1/120.22x   
32,768      0.0377      4.6372      a4ec6c3f    c0a7c7e9    1/123.04x   
65,536      0.0696      8.2590      e18ee207    f873c7c8    1/118.66x   
131,072     0.1301      16.5376     e949fd41    84608262    1/127.12x   
262,144     0.2605      30.7391     a8fe5730    2fb136ba    1/118.01x   
524,288     0.5144      58.4322     c7c95469    dd61a524    1/113.59x   
1,048,576   1.0611      117.1636    1fbbd4ff    6ed4af0a    1/110.42x   
2,097,152   2.1319      234.7411    81209a75    ff9f600f    1/110.11x 
```

New Code on GPU:
```
============================================================
PERFORMANCE COMPARISON
============================================================
Households  NumPy (s)   JAX (s)     NumPy Hash  JAX Hash    Speedup     
------------------------------------------------------------------------
32,767      0.0305      1.1938      2584ff76    9a97a0ef    1/39.20x    
32,768      0.0225      0.6582      a4ec6c3f    c0a7c7e9    1/29.29x    
65,536      0.0423      0.6841      e18ee207    f873c7c8    1/16.17x    
131,072     0.0790      0.7263      e949fd41    84608262    1/9.20x     
262,144     0.1510      0.7738      a8fe5730    2fb136ba    1/5.12x     
524,288     0.3497      0.8522      c7c95469    dd61a524    1/2.44x     
1,048,576   0.7123      0.9904      1fbbd4ff    6ed4af0a    1/1.39x     
2,097,152   1.3929      1.2859      81209a75    ff9f600f    1.08x 
```
I just ran the benchmarks on my GPU and the performance before the PR was indeed atrocious. After the changes we even start seeing a speedup on GPU compared to Numpy, which I assume will only get larger, the bigger the dataset. The Jax version also spends time compiling when first running a function, so this comparison might not be the fairest, depending on the use case.