@JuergenWiemers 2025-08-08
@hmgaudecker 
> 1. Add a test to TTSIM that makes the same error appear as in the GETTSIM test suite

I added a test that _would_ fail without the early return.:
```python
if p_id.shape[0] == 1:
    return
```

Without the early return, the test fails for the same reason that the GETTSIM tests failed: a one-row input dataset with `p_id==0` is falsely detected as a duplicate.

Is this the test you want?