@hmgaudecker 2025-08-08
> Finally fixed the silly CI failures. Forgot to convert the Pandas `p_id` to NumPy/JAX. 🙈
> 
> But why does
> 
> ```python
> p_id_sorted = xnp.sort(p_id)
> ```
> 
> locally pass all tests (in a JAX-enabled Pixi environment) while it must be
> 
> ```python
> p_id_sorted = xnp.sort(xnp.asarray(p_id))
> ```
> 
> to pass GitHub's CI?

Are you running `pixi run tests-jax` ? Else you'll still use numpy as the backend, even if jax is installed.