@hmgaudecker 2025-08-08
> Without the early return, the test fails for the same reason that the GETTSIM tests failed: a one-row input dataset with `p_id==0` is falsely detected as a duplicate.
> 
> Is this the test you thought about?

Precisely, many thanks! 

Basically, the situation we had should not happen: TTSIM test suite passes but GETTSIM test suite fails (essentially, GETTSIM tests should only be about policy contents, not about infrastructure). Given the recent split, this is to be expected though. However, we'll (almost) always want to create a TTSIM test then.